Hodgkin huxley
So i coded the hodgkin_huxley. Im pretty sure this is how you do it. Took formulas straight from the paper. So for the part of calculating potentials if you use a timed plot you just would multiply the change in time at each potential for each step. Didnt know how you coded it for yours so just left it there.